Update 102-fibonacci.c file

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int fib[50];
+	long int fib[50];
 	int i;
 
 	fib[0] = 1;
@@ -15,7 +15,7 @@ int main(void)
 	for (i = 2; i < 50; i++)
 		fib[i] = fib[i - 1] + fib[i - 2];
 	for (i = 0; i < 49; i++)
-		printf("%d, ", fib[i]);
-	printf("%d\n", fib[49]);
+		printf("%ld, ", fib[i]);
+	printf("%ld\n", fib[49]);
 	return (0);
 }
